+++ RELEASE_NOTES.md
@@ -1,5 +1,13 @@
 # Pushrod Releases
 
+## 0.1.5
+
+- Modified Base Widget to clear its invalidation state after a draw.
+- Set window `swap_buffers` flag to false when added.
+- Run loop optimization:
+  - Checks to see if widgets need to be drawn/refreshed (invalidated)
+  - Upon invalidation, draw is called, and display buffer is swapped
+
 ## 0.1.4
 
 - Run loop optimization: